Scan records are generated and served as typed messages
scan.proto defines the scan record model - TVSeries, TVSeason, MediaFile,
SidecarFile, TrickplayInfo, FileInfo and its stream parts, FileStat,
ImageInfo. internal/shared/scanmodel aliases all of them; ScanResult stays
a plain struct (a transient pipeline container, never stored or sent
field-for-field) holding *TVSeries / []*MediaFile. RecordType, DirectoryType,
SidecarType and SidecarCategory become free-string consts so they compare
directly against the plain string fields on the messages, matching how the
metadata slice handled Kind/Scope and how directory_scanner.proto already
models scan_type/category. Timestamps are google.protobuf.Timestamp,
converted at the filesystem edge with timestamppb.New / AsTime; the name
parser's ints are narrowed to int32.

LocalDirectoryService returns the generated messages instead of opaque
*_json bytes; services/local_directories.go no longer JSON-encodes anything,
and GetDirectoryNFO returns a typed Metadata (agentproto.NFOReadReply.Metadata
is *metadata.Metadata now). No UI consumes this service yet, so there is no
hand-written TypeScript mirror to delete.

Persistence moves to protojson: scanmodel.MarshalStored / UnmarshalStored
(proto field names, unpopulated fields omitted) back the Mongo repo, which
expands the JSON into a BSON subdocument so stored field names stay
snake_case and the indexes keep matching. Record identity is the collection
_id only - the message id field is stripped on write and stamped back from
_id on read; directory_id is stored as the directory's _id hex. Two
protojson quirks are repaired in replacementDocumentFrom: 64-bit integers
(size_bytes, the stat.* counts) are coerced from quoted strings back to
int64 via a descriptor walk so they stay numerically queryable, and
scanned_at is normalized to a fixed-width UTC layout so the stale sweep's
string "$lt" is chronological.

go build/vet/test green across mediascan, scanmodel, mongostore, services
and agentregistry; golangci-lint and buf lint clean; the agent no-mongo
boundary test is unchanged and passes.

+++ internal/agent/mediascan/imageinfo.go
@@ -44,7 +44,7 @@ func readImageInfo(path string) (*scanmodel.ImageInfo, error) {
 
 	return &scanmodel.ImageInfo{
 		Codec:  format,
-		Width:  config.Width,
-		Height: config.Height,
+		Width:  int32(config.Width),
+		Height: int32(config.Height),
 	}, nil
 }
